python/requirements.txt
fastapi
uvicorn
gunicorn
pyspark
delta-spark
